Add ability to load ifo data from YAML and MATLAB .mat files
This adds a Struct() class that mimics a MATLAB struct by storing values as
class attributes, but also includes methods for converting to/from dicts,
yaml, and mat_struct objects.

A convenience function load_ifo() is included to load an ifo definition
either from a .yaml or .mat file or from an included <ifo>.yaml.

An 'aLIGO.yaml' definition is included.