Beginings of a test suite
This utilizes matlab.engine to run MATLAB gwinc on it's default IFOModel.
The resultant data (ifo and noises) are saved to gwinc.mat file, which is
then loaded by the test command.  The loaded ifo parameters are used to
create a pygwinc ifo model, from which pygwinc noises are calculated.

The difference between the matgwinc and pygwinc noises are then plotted
and displayed.b