combine suspQuad and suspBQuad functions
These two functions were basically identical except for the material
used for the test mass stage, and a hacky reduction of the spring
constant for silicon.

suspQuad is modified to support different materials for the test mass
stage, and a hacky switch is added to adjust the spring constant.
suspBQuad is now just a wrapper around suspQuad specifying that the
test mass material is silicon instead of silica (there is definitely a
better way to do this, by e.g. specifying the fiber material in the
Stages struct list).