suspension: update suspQuad function from matgwinc supporting stage iteration
This is a rote copy of the new suspQuad code in matgwinc that iterates over
the stages, and supports tapered fibers:

https://git.ligo.org/gwinc/matgwinc/commit/a59f01192a1ffc1644a45be536e31bc0eb78bd7c#9cb6e594e14c1d61dbb3c91549166a2d48db194f

This allows for setting per-stage temperature and wire and blade properties.

This version differs slightly in some generalizations about the number of
stages, although it doesn't quite yet support an arbitrary numer.