Match matgwinc value of nu_small in precomp, and enable strict testing
With this tweak, most noises agree within 1 ppb.  Exceptions are
seismic and substrate TE (which still have open bugs in matgwinc).

Arguments added to gwinc.test for setting a tolerance level, and a
list of noises to skip in the comparison.

The CI pipeline will now fail when tests fail.