Koji's restructured quad suspension code
These changes improve the modeling of suspensions that have a
temperature gradient along the fibers.  Each stage is split into an
upper and a lower joint, which can have different material properties
assigned in the ifo structure.  The temperature of a joint is that of
the mass to which it is attached.  (However, temperature gradients are
not yet supported within the top or bottom stages.)  The average of
the upper and lower joint dissipation is used to compute the noise.