fix gwinc() to use Noise style labels in noises dictionary
This fixes a regression in the move to the new Budget structured. The keys
in the gwinc() noises output changed to be the name of the Noise class,
instead of the label, causing the matgwinc comparison test to fail.

This fixes the keys in the gwinc noises output to be the labels to maintain
consistency.