Fix an error in the coating brownian noise calculation
A discrepancy was found by Gabriele Vajente between the gwinc
CTN calculation and several other codes, especially in the case
of large Poisson ratio in the coating.

See: https://git.ligo.org/gabriele-vajente/thermal-noise-models

This commit resolves the discrepancy by fixing an apparent error
in the Yam et al calculation that was implemented in gwinc.