test look for upstream master by default, and minor fixes
This changes the test to look for the upstream master in the user
current config and use that as the test reference by default.  This
seems like a more sensible default in general, and it's easy enough to
fall back to HEAD if need be ('-g HEAD').

This also fixes some things from the last commit:

* fixes the MR CI pipeline
* adds better command line documentation to the test
* fixes command help
* fixes bug in test cache pruing