simplify testing by removing unnecessary reference
The assumption should really be that the HEAD of the repo is the
current noise reference.  There is therefore no need to store a
special reference hash for the tests.  This simplifies things quite a
bit, in particular by not requiring submitters to update a reference.

This commit drops the reference and modifies the test to just compare
the current code against HEAD by default.  The option to compare
against an arbitrary ref is preserved.

The CI is updated for these simplifications.  The "test" job is now
simple generation of the budgets.  The merge request approval job is
renamed 'review:noise_change_approval`, and it's logic remains mostly
the same: it compares the current budgets against the head of the MR
target branch, and requires the same MR approval to pass.  The overall
MR approval process therefor doesn't change.