Fix coating brownian so it agrees with Hong
The small remaining discrepancy between Hong and gwinc is due to
an error in gwinc's implementation of the Yam et al model.

The coefficient in front of the phase derivative `dcdp_z` is
supposed to be `nN` (as written correctly in Yam et al, eq 1),
not `nN/2` (as found in gwinc).

In Hong et al, the coefficient is `nN/2`, but that's because
their phases are defined as one-way instead of round-trip (see
Yam et al, eq A2).

After correcting the error, gwinc and Hong agree, as expected,
for the case of equal bulk and shear losses and no
photoelasticity.