Remove carrier density noise from pygwinc
Based on the new calculation with Debye screening from
arXiv:2003.05345, there appears to be no plausible scenario where
this noise will affect the design of GW interferometers.  Koji
estimated the noise is below 1e-28/rtHz for Voyager, as did Florian
Bruns (lead author of arXiv:2003.05345).  It's also negligible for
ET, as discussed in the paper.

Fixes #50